Changes to reflect more resource-driven orientation. See src/test/resources/application-test.properties and integration tests for examples
