+++ src/main/java/org/ohdsi/webapi/service/CohortDefinitionService.java
@@ -15,34 +15,15 @@ import org.ohdsi.sql.SqlTranslate;
 import org.ohdsi.webapi.helper.ResourceHelper;
 import org.ohdsi.webapi.model.Cohort;
 import org.ohdsi.webapi.model.CohortDefinition;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 
 /**
- * 
  * TODO replace all cohort_definition lookups with that service, once available
- *
  */
-@Path("/cohort/")
+@Path("/cohortdefinition/")
 @Component
-public class CohortService {
-    
-    @Value("${datasource.cdm.schema}")
-    private String cdmSchema;
-    
-    @Value("${datasource.ohdsi.schema}")
-    private String ohdsiSchema;
-    
-    @Value("${datasource.dialect}")
-    private String dialect;
-    
-    @Value("${datasource.dialect.source}")
-    private String sourceDialect;
-    
-    private final JdbcTemplate jdbcTemplate;
+public class CohortDefinitionService extends AbstractDAOService {
     
     private final RowMapper<Cohort> cohortMapper = new RowMapper<Cohort>() {
         
@@ -71,27 +52,21 @@ public class CohortService {
         }
     };
     
-    @Autowired
-    public CohortService(final JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-    
     /**
-     * Returns all cohort definitions in the CDM schema, to be replaced 
-     * once there is a cohort_definition service
+     * Returns all cohort definitions in the CDM schema, to be replaced once there is a
+     * cohort_definition service
      * 
      * @return
      */
     @GET
-    @Path("/cohort")
     @Produces(MediaType.APPLICATION_JSON)
     public List<CohortDefinition> getCohortDefinitionList() {
         
         String sql = ResourceHelper.GetResourceAsString("/resources/cohort/sql/getCohortDefinitions.sql");
-        sql = SqlRender.renderSql(sql, new String[] { "CDM_schema" }, new String[] { this.cdmSchema });
-        sql = SqlTranslate.translateSql(sql, this.sourceDialect, this.dialect);
+        sql = SqlRender.renderSql(sql, new String[] { "CDM_schema" }, new String[] { getCdmSchema() });
+        sql = SqlTranslate.translateSql(sql, getSourceDialect(), getDialect());
         
-        return this.jdbcTemplate.query(sql, this.cohortDefinitionMapper);
+        return getJdbcTemplate().query(sql, this.cohortDefinitionMapper);
     }
     
     /**
@@ -99,15 +74,15 @@ public class CohortService {
      * @return
      */
     @GET
-    @Path("/cohort/{id}")
+    @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
     public CohortDefinition getCohortDefinition(@PathParam("id") final int id) {
         
         String sql_statement = ResourceHelper.GetResourceAsString("/resources/cohort/sql/getCohortDefinitionsById.sql");
         sql_statement = SqlRender.renderSql(sql_statement, new String[] { "id", "CDM_schema" },
-            new String[] { String.valueOf(id), this.cdmSchema });
-        sql_statement = SqlTranslate.translateSql(sql_statement, this.sourceDialect, this.dialect);
+            new String[] { String.valueOf(id), getCdmSchema() });
+        sql_statement = SqlTranslate.translateSql(sql_statement, getSourceDialect(), getDialect());
         
-        return this.jdbcTemplate.queryForObject(sql_statement, this.cohortDefinitionMapper);
+        return getJdbcTemplate().queryForObject(sql_statement, this.cohortDefinitionMapper);
     }
 }
