+++ src/test/java/org/ohdsi/webapi/test/VocabularyServiceIT.java
@@ -3,27 +3,29 @@ package org.ohdsi.webapi.test;
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
+import org.ohdsi.webapi.vocabulary.Vocabulary;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.TestRestTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.client.RestTemplate;
 
 /**
  *
  */
 public class VocabularyServiceIT extends WebApiIT {
     
-    @Value("${vocabulary.service.endpoint}")
+    @Value("${vocabularyservice.endpoint.vocabularies}")
     private String endpoint;
     
-    private final RestTemplate restTemplate = new TestRestTemplate();
-    
     @Test
-    public void concept() {
-        log.info("Testing concept endpoint");
-        final ResponseEntity<String> entity = this.restTemplate.getForEntity(this.endpoint + "/concept/1", String.class);
+    public void vocabularies() {
+        log.info("Testing vocabulary endpoint");
+        final ResponseEntity<String> entity = getRestTemplate().getForEntity(this.endpoint, String.class);
         assertEquals(HttpStatus.OK, entity.getStatusCode());
         log.debug("Body:" + entity.getBody());
+        //or
+        Vocabulary[] vocabularies = getRestTemplate().getForObject(this.endpoint, Vocabulary[].class);
+        for (Vocabulary v : vocabularies) {
+            log.debug("Vocab: " + v.vocabularyName);
+        }
     }
 }
