+++ src/test/resources/application-test.properties
@@ -1,2 +1,9 @@
 baseUri=http://localhost:${server.port}${server.context-path}
-vocabulary.service.endpoint=${baseUri}/vocabulary
+#GET vocabularies
+vocabularyservice.endpoint.vocabularies=${baseUri}/vocabulary
+#GET domains
+domainservice.endpoint.domains=${baseUri}/domain
+#GET concept
+conceptservice.endpoint.concept=${baseUri}/concept/1
+#GET cohortdefinitions
+cohortdefinitionservice.endpoint.cohortdefinitions=${baseUri}/cohortdefinition
